KGRAG/requirements.txt
streamlit>=1.30.0
langchain>=0.2.6
langchain-core>=0.2.6
langchain-community>=0.2.6
langchain-openai>=0.1.8
python-dotenv>=1.0.0
openai>=1.12.0
neo4j>=5.17.0
pypdf>=4.0.0